Add project property to toggle copying live client graphics in build.
Buildserver currently uses the 'nmake' build to make use of the
'nocopyfiles' option preventing copying of live client graphics. Add
this as an option in the VS project file also, as the user-defined
property CopyLiveGraphics. Note that postbuild.bat is run in the
clientd3d project as opposed to when making resources in nmake due to
the limitations of makefile VS projects.