Add museum teleport scroll, historic portal limitations.
Added a new item, museum scroll, to facilitate transport to historic
cities. Purchaseable from Mychel for 4k, it will allow 20 minutes of
time in the historic cities before teleporting the user back to museum
(logged on or not). Mychel will buy back the scrolls if they are offered
to him, but they are otherwise undroppable. Using the scroll also
teleports the user back to museum (and does not destroy the scroll).

Added Mychel to questengine, fixed up spacing between portals in CN
museum.

Decreased the move timer delay after moving onto a teleporter in
classic client. Hopefully alleviates some of the issues with moving
onto portals.