Allow smaller GH chest doors to stay open longer.
Some GH chests are only accessible for a short period of time while a
door is open. Increase the length of time Jaarbas' GH chest door stays
open  to make using the chest less tedious, and bring it in line with
other similar guild halls. Change the code on all 3 to make setting this
easier if we need to change it.