Allow phasing in guild halls.
If phase expires, returns phased player to last safe spot (i.e. not the
hall), only gives penalty if user was recently in PvP.

If a user tries to phase back in inside the hall, if they are not
allowed in the hall they will be teleported back to the foyer to prevent
rushing in, waiting for reinforcements then phasing back in to allow
entry to others.