Add missing property printing avoiding duplicates.
Properties are often removed when adding or removing features from the
game, and the server prints every instance of an object found with a
missing property during startup. This can be quite excessive (10000s of
lines printed) when only one instance is really needed for each class.

This change saves each instance into a list which is then printed after
loading all objects.