Misc formatting/fixes.
Formatting in battler.kod, and removed a logged on check in
SetSecondWind (NewVigor checks for $ poOwner).

Formatting and code style in monster.kod.
Added daemon helm to frenzy sell list.
Changed the GetMaster call in the CreateTreasure loot check to
IsOwnedByPlayer to catch cases where a minion of a minion is doing
damage.

In brain.kod line 85/86 is incorrect - if NOT state & STATE_LIMBO will
negate state and then & that with STATE_LIMBO rather than check for the
presense of the STATE_LIMBO bit (old bug). Commented this code out since
it doesn't seem to be useful.
Added an IsClass check for Battler to SomethingAttacked to prevent doing
hatred/distance/etc checks if a room (pogg king EQ) or item (nerudite
armor) does damage to something.

Remove fog from ghost due to transparency drawing issues. Cap zombies
at 16 rather than create them indefinitely.

Lower lupogg's ranged attack radius to 6 row/col. Fixed the height of
the projectile used by lupogg and lich queen (was using same as that
for pogg king, which is too high).