Allow statues and log ghosts to display headgear.
Masks, helms and circlets were not being shown on the statues in Hall of
Heroes, or on logoff ghosts. These items need to be handled specially in
both cases because of how they modify the character's head/face (setting
some things blank etc) and because not all helm-slot items will remove
hair.

Translations are saved for ghosts but not for statues, as colored items
won't show up on statues anyway.