Fix glowing lutes not showing light; light level reset.
Lutes enchanted with the glow attribute increased the light level of the
wielder, but did not show a dynamic lighting effect to others in the
room. Now fixed for players and human mobs.

Fixed glowing lute light level and touch spell light levels not being
taken into account when a player's light level was being recalculated.