Don't display two phased in messages for pens.
If a player runs out of phase time and pens, currently both the teleport
and the 'phased in' messages are displayed to the room. Prevent the
'phased in' message from being displayed if the phase in is due to a
penalty (i.e. the player is now at max phase time remaining).