Fix client audio bugs.
Client was occasionally crashing if a user tried to set their music
volume immediately upon logging in, likely because the intro music
filename pointer was bad (passed in from intro.dll but this module gets
unloaded) and this pointer was in lastmusicname. This field is now a
preallocated array, set via strcpy.

The order and calls performed when the user set any music/sound
preferences was not ideal - UserToggleMusic was called every time the
preferences were set, potentially resetting music when it shouldn't, and
also potentially setting it to play the wrong song (from lastmusicname).
Volume was set after the music toggle was checked, when the other way
around is likely better (same with sound).

Also now keep track of the music filename to play, even if music is
switched off. This means if the user decides to switch music on, they
get it straight away instead of only after an update of the music to
play from the server. Sounds need to be updated to do this at some
point.

Lowered the default volume of music/sound from 100 to 75.