Add 3 call opcodes to handle non-settings calls.
There are two types of C calls - calls with settings (named parameters
in Send, Create etc) and calls with no settings (First, Nth, Length
etc). The majority of calls won't have any settings (even Send calls
with no parameter setting fall into this category) but all calls
currently get the byte for the number of settings and check if it is 0
to skip getting setting data.

This change adds 3 opcodes to deal with calls that don't have any
settings, to skip the 0 assignment and check.

When testing a complicated message (FindPlayersOverLearnCount with live
data) runs ~10% faster.