Fix up hold effect on players when logging in unsafely.
LogonDelay handles players logging in or unphasing while unsafe, and
sends EFFECT_PARALYZE_ON to the client but does not actually hold the
player server-side.

Switched this for casting hold on the player with the login delay time
as the spell duration (login timer no longer necessary). Hold state
(report) is set to 0 so hold-specific messages are not used, to
differentiate this hold from a properly-cast one. Login delay hold time
is reset if the user logs or phases again while under the effects of the
hold.

Moved the LogonDelay message call to before the user is actually logged
on, as casting a spell on the user after this point can cause issues
(multiple enchantment icons).