Add new PE anti-magic aura spell.
Converted AMA to a personal enchantment. Room enchantment still exists
as 'anti-magic room' but is unavailable (players automatically get the
PE version). Self-cast only, causes all spells cast by the enchanted
player to take a spellpower hit, and any single-target spells cast on
an enchanted player also take the spellpower hit. Spellpower is
subtracted after the spell has succeeded but before the effects are
applied, so fizzle rate is not increased but the outcome of the spell
(e.g. hold duration) is.