Add support for followers doing non-LOS teleporting.
Followers need to be teleported to their master when the master casts
blink. Currently, they won't be teleported if there is no line of sight
to the new position. Fix this by setting mob_override to TRUE in the
call to Teleport from blink.kod, and add this to various Teleport
implementations. Tos arenas already support mob teleports correctly,
although followers cannot enter those rooms.

Changed the while loops in UtilGoNearSquare to for loops to save a few
lines of code, plus general cleanup/code style.