Add msg stubs to room.kod, perf changes in monsroom.kod.
Room needs GetPatrolRadius and GetLeashRadius stubs to handle mobs that
get seduced and taken to another room (that may not be a monster room).

Created a combined CountMonsters and CountHeroes in monsroom.kod
(UpdateBattlerCount) to be used when both properties need to be updated.
This however isn't called (just there for convenience) as the
hero/monster counts are now updated directly in NewHold and LeaveHold
for performance.