Handle list node array resizing in SendList functions.
If the list node array is resized, any existing variables pointing to
the old list node memory could be invalid. To handle cases where this
happens during a call to SendBlakodMessage inside a SendList-family
function, list node ID needs to be saved before the call and the list
node obtained again afterwards.