Standardise wall durations in milliseconds.
Some issues with various usages of wall spells/elements in seconds or
milliseconds. Standardised by having any call to a wall element's
Constructor assume duration is already in milliseconds.

The main issue here was an integer overflow when casting sporeburst -
the spell sent the duration in milliseconds, then the active wall
element assumed duration was in seconds and multiplied by 1000 to get
milliseconds, ending up with a negative duration (bound to 1000) and
causing sporeburst to have no effect.