Add C call to get unix timestamp in a kod string.
Since unix epoch time exceeds the value of a kod integer (which can only
use 27 bits for a number, not including sign bit), to correctly express
unix time in seconds it needs to be stored as a kod string. This
function returns the time as a string (GetTime() without the offset).
Will be used for the time value in posts and possibly mails, to make
them future-proof.