Add temporary C call to convert old offset timestamps.
Used for cases such as newsglobe post dates where a kod integer
timestamp in seconds, with offset, needs to be converted to a proper
string unix timestamp that would overflow the kod int data type. The
time offset itself is too large to fit in the kod int data type, so the
conversion must be done in a C call.