Mail and newsglobes converted to use unix timestamp.
To fix the issue of needing a time offset in the clients (due to kod int
overflow) both mails and newsglobes will keep track of unix time in
addition to the offset time returned by GetTime().

The unix timestamp is kept as a string, as this is the easiest way to
store a number larger than MAX_KOD_INT in a single data field. This is a
temporary measure until kod can support larger integers, however this
string is converted to an integer before sending to client so the client
just needs to handle a 32-bit int.

Calculations such as recent post time and deleting old mails are done
using the short timestamp.