Disable AE imp penalty in Seafarer's Peak.
Added a property to the OutdoorsA5 class to toggle whether any AE spells
cast have a penalised improve rate. Allows players on sacred haven
servers to have their spells discorded while not suffering an
improvement penalty. Not expected to cause issues on PvP servers, but
can be toggled back on if necessary.