Fix activatable quest bugs.
Rare bug with the activatable quests (guild shield blacksmith part,
'good' path of Jala pupil quest). These two quests are assigned specific
restrictions based on the results of another quest.

If the quests are assigned restrictions but never used, they are not
deleted and stay in the active quests list until the next quest system
recreate. Once the restrictions are added, the quests can't be assigned
to another player (i.e. set of restrictions) and since the quest system
can't assign more than the max number of active quests for each quest
template, eventually there will be no more available quests to activate.

To solve this, a new quest status has been added,
QN_STATUS_WAITING_FOR_PLAYERS_TIMED. This status goes into effect if the
initial quest has a time limit, allowing the bugged quests to be deleted
if not used in the allocated time. Activated quests with infinite time
can still be created by specifying 0 for the timelimit.