ET/phase/seduce interaction fixes.
If a player has ET cast on them and phases, the ET should also phase and
not move (and can be walked through). If the ET is seduced by a
different player, it should stay unphased as long as master and target
are online.

Seduced ETs should teleport to the new master (rather than the original
target) if the target logs off/on.

Added some new object/drawing flags messages to User to override phase
and spectate's effects so that ETs that have a phased original but
should not themselves be phased behave/look correct.