Fix cancelling RE issue with multiple REs.
If two players in range of each other have the same RE up and one of
them cancels it by recasting the same spell, they did not immediately
come under the effects of the other player's RE.

Added a new event, EVENT_CANCEL, and added a RE recheck for it in
BreakTrance. An existing event could be used, but this seems more
correct. Also check for EVENT_STEER anyway.