Add a workaround for PvP heal issues.
Non-harmful spells will now check if the caster can fight any of the
battlers attacking the target of the spell. What is specifically checked
is whether any Battler (player/mob) entry in the targeted player's
plHurtMeRecently can attack the player attempting to cast the spell, in
the room occupied by the target.

If any Battler can attack the caster, the spell will still be successful
to deal with cases where multiple PvP-flagged subsets of players are
fighting.