Fix apparition spawning/attack/kill issues.
Apparitions should not be allies to anything except the master.

Place apparitions near the target to make them easier to see, but place
them on the target as a fallback if random position can't be obtained.
Uses the new C call for getting a random point (which can possibly place
things too close to walls).

Player kills by apparition will now be correctly attributed to the
creator of the apparition.

Apparitions that have been captured with seduce will now be deleted
when seduce wears off.