Disable debug printing for FoL/KF.
Attempting to remove the attack modifier for these two spells is
somewhat unavoidable without special case logic somewhere, as they can
be blanket-removed from all players in a room when they do not affect
all players. The special case is placed where it should have the least
effect, supressing the debug message for these two cases.

For reference, KF can be double-removed if a phased player is checked
when unenchanting all occupants (but we want to check for safety). FoL
can be double-removed if a player wasn't enchanted due to having
negative karma (but we want to check in case their karma dropped after
gaining the enchantment).