Fixed some bugs with name filtering.
Debug strings (i.e. raw "text" in code) can't be stored in lists,
tables, properties or classvars because when the game/code is reloaded,
the references are not updated (i.e. same issue as message IDs). As a
result, the IDs on a live server won't match up to the expected string
value after some changes to system.kod.

Removed some of the reserved names, I think it's unlikely this will ever
be useful for the majority of names on the list (and some are fairly
generic).