GetObjectByID, class name changes in server code.
Added GetClassNameByID() and GetClassNameByObjectID() functions to
obtain a class name while never returning NULL. GetClassNameByID() is
used in GetObjectByID() but is inlined (GetObjectByID() is never
inlined).

Added GetObjectByIDInterp() to skip the class name retrieval altogether,
causes this function to be inlined in interpreter functions but not in
RetrieveValue(), possibly because RetrieveValue is already an inlined
function or it crosses some threshold for complexity disallowing both
inlinings to occur.

Increased safety a little in the goto instructions.