Convert First() C call to opcode.
First() is the most used C call after Send(), and is simple enough to
function as an opcode vs a 'library call'. Converted to an opcode, a
First() instruction storing into a property is ~45% faster than the C
call, and storing into a local is ~48% faster. This also has the side
effect of making all foreach loops faster, as they must call First each
iteration.

Added verbose error reporting to the compiler, some error messages seem
to have a large improvement (listing the expected input) but some might
be confusing due to terminology.