Add 'register callback' dialog to server UI.
Adds a toolbar icon to the server UI next to the MOTD entry dialog,
with a 'new page' icon. This opens a menu where details for a 'kod
callback' can be set up, which are sent to the RealTime object to be
added to the callback list when the 'OK' button is pressed. User can
enter object/message to call, time (y-m-d-h-m-s) to activate the
message, and up to 4 optional parameters. Error checking on the inputs
should be extensive enough to catch any errors, and a help menu with
more details is included.

Example usage is setting a frenzy at a given time (no params required,
just time fields, object 0 and message StartChaosNight).