UDP error/debug printing changes.
Don't print a debug message for lost/malformed UDP packet, as dropped
packets seems to happen a small % of time (which adds up to a lot of
logged messages).

Move the 'unknown session or hangup session' error to debug log, this
one can be caused by a client disconnected due to lag that is still
trying to send packets (possibly don't need to log at all).