Add throwable snowball weapon, snowball dispenser.
Add a new class of weapons that can be thrown with one example, a
snowball. ThrownWeapons are a combination of the NumberItem and Ranged -
they consume items from a numbered stack on use, and will display a
projectile. They currently use the fire/arch skill infrastructure, but
should be moved to new skills when they become available.

As they are similar to NumberItems they can also have dispensers, which
for the snowball is a snowpile.

Set snowballs to 0 value, 0 damage and bypass safety/faction/room
checks. They can be set to work as real weapons by toggling the
vbDoesDamage property.

Add an ATCK_WEAP_COLD type to handle snowball dmg type.