Use UTC time to set in-game time.
The values used to choose the current in-game time are somewhat
arbitrary now and hard to sync up with client's internal time. Using UTC
time to set game time means this is easily reproducible on any client at
any time.

Ideally the server would sync time with the clients to allow for changes
in server time, but as long as server time isn't adjusted manually
clients will always display the correct internal time.