Get client settings on login.
The amount of memory sent by the client when logging into the server is
generally invalid now, as the field is 32-bit (signed) and most users
have more than 2gb ram. Use this field instead for a set of client flags
detailing which settings are turned on, more out of curiosity than
anything else.

Currently checks for sw/hw renderer, large/small graphics, music on/off,
dynamic lighting on/off and particle effects on/off. Data is logged in
log.txt as an int value (if we want to grab it via log parsing) and as
readable text for each field. These flags are only checked for classic
client.