Fix client losing connection due to lost D3D device.
Current code for detecting and resetting a lost DirectX 3D (D3D) device
handles it by running a loop calling
IDirect3DDevice9_TestCooperativeLevel to get the device status. This
works, but a side effect is the client won't send/receive data while
this is happening (including pinging the server). This means the client
will get disconnected as inactive after 6 seconds.

Change the loop to a test before beginning each frame, returning if no
device is found.