Add admin 'show post' command.
Command 'show post' works by taking a globe ID/index, and a post
ID/index, and retrieving the specified post via kod calls. Prints the
details of the post to admin window.

This method was chosen for ease of data handling outside of the game
(i.e. for use by an IRC bot).