Add ChangeSectorFlagBSP C call, depth/height fixes.
New C call ChangeSectorFlagBSP allows changing sector flags server-side
live. Main use is to update depth flags when the height of water changes
(i.e. turns to ice) so that height calculations are accurate.

Can also be used to alter SF_NOMOVE status provided the sector is
tagged. We could add a method of tagging servers by location, or
altering sector flags by location but that is beyond the scope of this
commit.