Handle integer overflow in player food consumption.
After piTimeLastStomachUpdate is reset to 0 (e.g. timebug fix) or if a
player has not logged on for 129 days, a calculation involved in whether
they can eat will overflow and cause them to be unable to eat for a
further ~15 min. If an overflow is detected, let them eat straight away.