Bound slope height checks to min/max possible.
Client move checking near passthrough walls will use the height of the
adjacent sector to determine whether the user can move into it, however
some of the x, y coords used in the move's intermediate steps are in the
original sector. For non-sloped sectors, just a height is returned with
no x,y checking but sloped sectors require x, y to return height. Using
x, y in a different sector (the original, not the sloped one) can return
a height value which is not actually present in the sloped sector, since
the height at x, y is just extrapolated from the slope equation.

To solve this, the height checks now use the x, y position at the
intersection of the move line and the wall line (code from Ogre client).