Allow qor priestess to hear players farther away.
Due to the way Qor temple is laid out, users tend to stand too far away
from Zuxana and assume she is broken (or that quests in general are
broken) when she does not reply. Allow her to hear everything on the
Longer term we probably want to provide more feedback to users (e.g.
check quests etc no matter where they stand and tell them where to
move) but short term this fixes one of the biggest 'issues' in the
quest system.