Don't check quest system for NPCs.
Only player speech should trigger quest checking, not NPC speech.

Note for future: if NPC speech (e.g. follower) WAS supposed to trigger a
quest, this would be the place to handle which NPCs could be heard (but
quest system has a bunch of player-specific checks which would all give
errors).