Save user's IP address for honor PvP system check.
Players can lose honor points in the honor PvP system if they log off
while fighting, however to avoid gaming the system it checks that both
players aren't using the same IP (not great for multiple users with same
IP...). This breaks down because the IP for a logged user wasn't
actually available, so this data needs to be saved. Get the IP on login
and use that for any call to GetIP/GetIPString.