Fix error in rscmerge.
Something changed in a recent version of either VS 2017 or the windows
SDK which causes recently compiled rscmerge to sometimes fail to find
any files. Specifically, it does have all the files in the arguments
passed but the null termination is incorrect (or some other string issue
which looks similar).

For now I've copied the rsc-handling code from the server, which does
work.