Fix Kardde's Canyon ledge, King's Way2 name change.
Kardde's Canyon ledge has a width below what we generally would use for
players/mobs - unofficial limit is 48 fine units, current width is 32.
With the movement changes in the client, it's thin enough that users
moving too quickly will hit one of the side walls of the next ledge up
and be knocked off the ledge. Widened the ledge to 48 fine units to fix
this (slight adjustments around ledge to accomodate new size).

Renamed the King's Way nearest Cor Noth to Outskirts of Cor Noth, both
to remove a duplicate screen name and because I need a resource change
so I can hotfix the above room.