Fix rare movement issue.
If the client is in SW renderer mode or in map mode in HW renderer, the
player can move into some walls if they hit it at the right angle. As
far as I can tell this is a consequence of the client being able to make
lots of small moves due to the low time needed to render a frame in those
rendering modes. Once they get close enough (min_distance) to the wall,
the 'slide along wall code' picks an angle/move which is just right to
gradually move inside it where the server then bounces them repeatedly.

Checking any move with an end inside min_distance works to fix it, but
presumably that will also create situations where users can be stuck
inside a wall and be unable to move out.

Added another fix for wall side issues in the SW drawing code to pick
the correct sidedef rather than handling NULL later on (but left the
other check in).