Fix for phase/room ench interaction.
Users phasing in after phase has run out of time will be teleported to
safety, however as the teleport is done last the user first has all
present room enchantments activate on them. Heat, if present, will tick
one round of damage on the user prior to them being teleported and can
result in the user's death. To fix this, room enchantments won't be
activated at all for a user phasing in just to teleport. Same applies
to radius enchantments.

Reordered the speech/trance-break handling for distil, so users can make
a typo and not waste their ruby.