Set regen flags whenever starting a regen/decay timer.
Currently when switching from regen to decay (e.g. taking a circlet off
over max mana) the player's mana can either snap back to max in 1 tick
(via the bound in ManaRegen) or tick down gradually (via ManaDecay).
This outcome depends on the time remaining on the current ManaRegen
timer - if more time remains than the new tick time, the timer switches
to ManaDecay, otherwise ManaRegen is called and the mana is reset to max
mana.

I think this likely isn't wanted, and seems to be due to missing regen
flag setting on any call to ManaRegen/ManaDecay (and matching calls for
health/vigor). Adding these flags makes it so that taking a circlet off
always switches the timer to ManaDecay.