Handle poObject in brain/monster.kod
Monster will continue to move towards the trigger object even if it has been picked up by another player or monster.
Trigger object must be in the same room, or in the hands of a battler that is in the same room as the mob. A battler can hold the trigger object, and exit the room and come back with the mob still interested in the object. If the item is dropped in another location the mob is no longer interested.
If something has left the room and mob is interested in an object. If the object no longer has an owner. The mob will forget about it.