Spawn reflections/ET away from target/caster.
Don't spawn reflections and evil twins directly on top of
casters/targets as this can generate an illegal move if the player
attempts to move prior to the mob moving away. Will attempt to spawn the
mob within 3 row/col, and defaults to the old method (on top) if random
placement fails.