Add hometown speech trigger.
Add hometown speech restriction constants for use in determining which
speech response to use based on the hometown of the user speaking. The
speechlib entry can specify a single town to give the response to, or to
give a response to anyone not from that town (i.e. response1 to marion
citizens, response2 to non-marion citizens). If more than one single
town is listed, all towns would need to be handled.