Active skills framework.
Works the same as casting spells, except using the 'perform' chat
command. Performing a skill sends a BP_REQ_PERFORM, analogous to
BP_REQ_CAST, to the server (which just logs the skill name
currently).

BP_SKILLS protocol has three fields added - the number of targets,
the school num and whether or not it is an active skill.