Fix SomethingMoved wallelem checks, meteor fix.
Wall elements and bramble need to use an intersection test of the move
with their position to correctly affect battlers who cross the element.
The main scenario this addresses is the case where a user moves across
an element, but does not send any move requests that coincide with the
element's range. This could be accidental (e.g. lag) or deliberate (~10
lines of code in classic to never send a request on a wall element).

Fixed meteor not checking height - as it was using a line of sight call,
it could go places it could see but not really access. The fix now means
it can't sometimes go places it can theoretically 'fly' to because the
room geometry blocks it. This should result in less 'unwanted' scenarios
though.

Fixed up a bit of formatting, and added a setting for whether to delete
walls around battlers on entry (default on, same as live).