Prevent spawning multiple door timers in oc06.
Konima's Ascension door had no check on whether it was already open
before spawning another timer to close the door. This meant multiple
timers could be created, all but the final one not attached to a
property. Causes error spam on recreate if an unattached timer closes
the door on the now deleted room.