Fix one of the 'can't attack' bugs.
Encountered sometimes when trying to cast a non-damaging spell affecting
a monster. Caused by the 'spell help check' which checks the
plHurtMeRecently list to see if the attacker can attack targets on that
list. Useful for pvp to avoid interference, but doesn't need to be
checked at all for monsters.