Zero GetString() string buffer before loading string.
Junk data can stay in this string buffer, and some functions calling
GetString() use it in a less than safe way. Ideally the offending code
should be refactored (i.e. alias loading) so it isn't handling the
result of GetString() this way, but zeroing the string buffers used is
also a safe addition regardless.

Fixes a very rare client crash while loading aliases depending on what
was in the string buffer prior to its use in CmdAliasInit().