Add QN_TYPE_MONSTER_MULTIPLE quest type.
- QN_TYPE_MONSTER_MULTIPLE is like QN_TYPE_MONSTER but requires multiple
mobs of that type to be killed to complete the quest.
- Format for monster_list elements in quest template is [mob_class,
num_to_kill], and multiple elements of this type can be included in one
monster_list.
- Added %NUMTOKILL quest text modifier for the QuestNode to insert that
variable into quest strings for NPC speech.
- QuestNode keeps track of # mobs killed and target #, and each can be
accesed e.g. by QuestTemplates for descriptions.
- Improved the handling of QN_TYPE_MONSTER on recreateall, users will
now need to kill the same mob as before the recreate vs random mob from
that quest.