Cleave weapons wings adjust, AOE XP fix.
- Make cleave only add one weapon swing to piWeaponSwings on using the
skill, instead of one swing for each target hit.
- Adjust handling of gain flag resetting when killing mobs with an AOE
(cleave, ice nova, EQ). Gain flags shouldn't be reset on target
switching or target kill in the middle of performing an AOE attack on a
group of monsters. Gain flags are instead reset at the end, if all
targets were killed.