Chest recreate adjustment, globalgive fix.
- Prevent empty char slot user objects from receiving items from
GlobalGive.
- During recreateall, existing chests are saved and replaced after room
recreate, but this requires the new chest to be in exactly the same
place as the old one. If we move a chest during an update, this system
breaks. Changed chest recreate so it no longer deletes the old chests if
they can't be placed, and logs a debug message so we can manually fix
it.