Misc small fixes.
- Adjust entry angle going into Berg Merchant since the room's
orientation is slightly different to the door placement in Berg city.
- Prevent weapons which have a second attack (multishot bow) from
counting as a second weapon swing.
- Prevent empty char slots from receiving spells/skills via
AddSpell/AddSkill (e.g. if this is called on every player to
give a new utility spell, the uncreated players would get it too).
- Small cleanup of BergMerchant, and set the plFor_Sale list to have 4
elements in case this NPC gets a conditional item later.
- Set upgrades list to $ in guild decorator on deletion.
- Add error messages to SendList functions if trying to send to an
invalid object.