Allow minions+follower to follow through doors.
- Intra-room doors (CV, necro lair) currently block minions and
followers from following the player. Minions handle this by teleporting
to the master after a set period. Handle these doors by instead having
follower + minions send a SomethingTryGo if the player successfully uses
an intra-room door.
- Move IsMonsterQuestEquivalent from Monster to Battler. Follower
currently calls this on room objects which might include other players.
Rationale for moving it and not excluding players is that eventually we
might want a quest where another player is the target of a quest.