Fix some small follower-related bugs.
- Shrunken head mask has some issues when worn by followers - because
they drop the item as the unuse it, when offered a new headgear item to
wear it led to DefenseModifier.ReqUseSomething being called on the
owning room because ShrunkenHeadMask.ReqUseSomething dropped the item.
This also caused the newly offered item to fail to be equipped.
Returning TRUE from ShrunkenHeadMask.ReqUseSomething fixes this by
dropping the shrunken head mask and allowing the new one to be equipped
due to the TRUE return. ShrunkenHeadMask.ReqUse also needed a check for
&Human, so that they would unequip circlets when equipping the mask.
- Martyr's grace needed some additional checks to work on followers.