Remove grey background on desc bmps, others.
Description dialog box images all have a light-grey background, and use
light grey pixels wherever the image has transparency. This is caused by
using a light-grey system 'brush' for drawing bitmaps, i.e. any
transparent part of the bitmap will receive the grey color.

I've fixed this with two changes:
1. Create and use a brush which has the Meridian transparent color
(cyan) as the background for anything that will be displayed in a dialog
box.
2. Use the TransparentBlt GDI function instead of BitBlt and StretchBlt,
as TransparentBlt allows specifying a transparent color. Specifying cyan
for this allows removal of the image background in dialog boxes. Note
that I considered using the system brush background color as the
transparent color, but couldn't guarentee it wouldn't somehow be used in
an image and thus cause transparency where it is unwanted.