Fix flag recreate issues, add recalc for guild flags.
Original flag recreate mechanism was broken because setting a flag's
faction via SetFaction calls back to TerritoryGame.FlagClaimed() which
recalculates the saved flag list. SetFaction is called when placing new
flagpoles and also when setting the flag's faction from the saved flag
list, and both of these actions will delete the saved flag list.
A fix was added for this which operates in addition to the above
(broken) method; the fix just saves a deep copy of the saved flag list &
replaces flag faction after TerritoryGame.Recreate() has run using a
custom SetFaction() which doesn't call back to TerritoryGame. This is
inefficient because now there are two flag recreate mechanisms running.

To fix this properly, the flag recreate fix is removed and a pbRecreate
property added to TerritoryGame which is set TRUE while recreate is
running. If set, TerritoryGame.FlagClaimed() won't reset the saved flag
list.

- Also added missing guild & capturer fields to saved flag list so they
can be replaced properly.
- Added message to recalc guild territory list.
- Added a temporary message to fix flags prior to the recreateall, as the
current live flagpoles don't have a poGuild property.