Fix trance/cleave issues.
Cleave has an unbreakable trance period, but all calls to BreakTrance
assume the trance is broken when the call is made. This means it is
possible to cast additional spells (or attack) in the period between
starting a 'perform cleave' and the cleave actually being performed.

One obvious option to fix this is to have BreakTrace return FALSE if the
trance can't be broken, and use that result to disallow an action. I've
opted not to do this because it's a bit more intensive (calling
BreakTrance on all enchantments the player has) and might lead to
unwanted effects if a player has several 'trance-like' enchantments,
some of which get broken and some which don't, which prevents the action
anyway.

Instead, Player now has a CanBreakTrance message which returns TRUE if
their trance can be broken, and FALSE if it can't. Checks Trance object
(if present) and spell trances by checking the return of CanBreakTrance
on them (checks ability object for Trance). Anything that shouldn't
override a trance will call this (casting another spell, attacking,
performing, using/unusing an item) but some things don't need to check
(changing rooms) since they should always succeed (and changing rooms
fails cleave since the targets aren't valid anyway).