Slight refactoring to remove some class sends.
The Send C call has a feature where it can send a message to all
instances of a class (e.g.
Send(&RainbowFern,@ChangeColor,#iHour=piHour);). This is really useful
for directed messages where we don't have a list of objects handy or
where the alternative (keeping a timer on each rainbow fern) is also
expensive.

However 105 now has over 200,000 objects, and calling send on a class
requires checking if every object is that class (or if it is a parent
class of that object) which is slow on 200k objects. The NewGameDay
message in system.kod had three of these calls, and this message is
frequently the slowest-running message on 105 (likely when it also calls
NewGameDay, which has more class sends).

To fix this, I've added a NewGameHour message to Reagent (parent of
RainbowFern and TwinkleCap) so that even though one class send remains,
it can target both of these at once instead of calling ChangeColor
directly on both classes. For the PlantSoil NewGameHour class send,
System now keeps a list of all plant soils in the world (since there
aren't many) and makes the call on this list instead.