Fixes for gh public chest recreate, thrown weapon bug.
- Guild hall public chests are saved on recreate like other chests, but
not correctly reassigned to poPublicChest when replaced. To fix,
intercept NewHold calls where the object being placed is a chest in the
public chest coordinates and assign it (don't need to delete existing
since that is done by system).
- Thrown weapons were Posting Delete in SubtractNumber rather than Send
(as in NumberItem). This worked for all cases except when placing a
stack of snowballs in a chest. If placing the entire stack in the chest,
it was deleted entirely if it didn't fit due to the ordering of the
calls. Sending this means it is deleted earlier and a new one returned
to the player.
- Post SubtractNumber when consuming thrown weapons so the weapon is
still around for hit messages etc. This removes the need to post Delete
in SubtractNumber itself.