Add monster flag bitfield.
- Convert the monster properties pbIllusion, vbSummoned, vbAnimated,
pbEvent, pbDontDispose and piEnch_flags into a bitfield piMonsterFlags
to save memory/savegame space.
- Move battler flag BFLAG_HARBINGER_MOB to monster flags.
- Add overrides for object.kod messages IsSummoned and
CountsTowardsSummonLimit, as vbSummoned is an object classvar (used only
in mobs and wall elements),
- Convert room.kod's CountMonstersWithBattlerFlag to
CountMonstersWithMonsterFlag.
- Change TriggeredInvasion's IsEventSpawn check on mobs for adding
rewards to a check for the harbinger mob monster flag, because faction
troops are also classed as event mobs and would add to the harbinger
reward total.